+++ src/db/mod.rs
@@ -7,5 +7,6 @@ pub(crate) mod deployment;
 pub mod marketplace;
 pub mod product;
 pub mod project;
+pub mod project_app;
 pub mod rating;
 pub(crate) mod server;